refactor: Refactored re-indexing logic
- Moved the `album-fracturization` adjustment logic into its on file as its pretty sizable.
- Simplified logic for `album-fracturization`
- Updated `getAlbumKey()` to apply `encodeURIComponent` to the `album` & `albumArtist` values.
  - Shouldn't affect functionality, but will improve reliability.
